requirements.txt
requests==2.32.3
tenacity==8.5.0
pydantic==2.8.2
PyYAML==6.0.2
